requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
python-dotenv==1.0.1
coreapi==2.3.3
uritemplate==4.1.1 